+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
 <!-- <START NEW CHANGELOG ENTRY> -->
-
+- Upgrade jupyter-extension dependency
 <!-- <END NEW CHANGELOG ENTRY> -->